separate create client, update test
