+++ fargate/upload-move/pkg/multipartUploader.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	s3types "github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"log"
@@ -222,32 +221,4 @@ func worker(ctx context.Context, svc *s3.Client, wg *sync.WaitGroup, workerId in
 
 	return nil
 
-}
-
-// CreateDefaultClient creates a client for the appropriate region data is being copied to
-func CreateDefaultClient(storageBucket string) (*s3.Client, string, error) {
-
-	// Get region
-	region := GetRegion(storageBucket)
-	if region.RegionCode == "" {
-		return nil, "", errors.New("could not determine region code")
-	}
-	log.Printf("Using s3 client for region: %s\n", region.RegionCode)
-	cfg, err := config.LoadDefaultConfig(context.Background(), config.WithRegion(region.RegionCode))
-	if err != nil {
-		log.Fatalf("Unable to load AWS config: %v", err)
-	}
-	regionalS3Client := s3.NewFromConfig(cfg)
-
-	return regionalS3Client, region.RegionCode, nil
-}
-
-// GetRegion from bucket naming scheme format gets the region name from the shortname
-func GetRegion(storageBucket string) AWSRegions {
-	bucketNameTokens := strings.Split(storageBucket, "-")
-	shortname := strings.ToLower(bucketNameTokens[len(bucketNameTokens)-1])
-
-	region := Regions[shortname]
-
-	return region
 }
